+++ CHANGELOG.md
@@ -799,7 +799,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - API
   -- Support for the new HMIP Cloud Update
 
-[unreleased]: https://github.com/hahn-th/homematicip-rest-api/compare/2.0.1..master
+[unreleased]: https://github.com/hahn-th/homematicip-rest-api/compare/2.0.3..master
+[2.0.1]: https://github.com/hahn-th/homematicip-rest-api/compare/2.0.2...2.0.3
+[2.0.1]: https://github.com/hahn-th/homematicip-rest-api/compare/2.0.1...2.0.2
 [2.0.1]: https://github.com/hahn-th/homematicip-rest-api/compare/2.0.0...2.0.1
 [2.0.0]: https://github.com/hahn-th/homematicip-rest-api/compare/1.1.7...2.0.0
 [1.1.7]: https://github.com/hahn-th/homematicip-rest-api/compare/1.1.6...1.1.7